[INTERNAL] Reader Collections: Allow for empty reader lists
If no readers are provided, no error is thrown and no resources are
returned.

This behavior was mostly already in place (and relied upon) but never
fully specified.

This makes it easy to implement tasks that just want to fetch any
available dependencies, even if none are available (for example if the
project has no dependencies).

See also https://github.com/SAP/ui5-fs/pull/455 for the opposite
approach of dissallowing collections without readers.

Also add tests for ReaderCollectionPrioritized (99% copy of the
ReaderCollection tests)