[INTERNAL] Resource: Do not flag as modified during creation
This is a follow-up of https://github.com/SAP/ui5-fs/pull/472

The setBuffer/setString calls in the constructor lead to the
sourceMetadata.contentModified flag being set to true, even though the
content has not been modified. To resolve this, use the internal
setter, which does not set any modification flags.